requirements.txt
autogen-agentchat>=0.2.0
autogen-core>=0.2.0
autogen-ext>=0.2.0
streamlit>=1.28.0
arxiv>=1.4.0
python-dotenv>=1.0.0
autogen-ext[openai]>=0.2.0
openai>=1.0.0
nest_asyncio>=1.5.0
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-mock>=3.11.0